backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Database (SQLite + PostgreSQL support)
sqlalchemy>=2.0.27
aiosqlite>=0.19.0
asyncpg>=0.29.0
alembic>=1.13.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# HTTP
httpx>=0.26.0

# Google OAuth
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.116.0

# AI/ML
anthropic>=0.18.0
openai>=1.12.0
langgraph>=0.0.26
langchain>=0.1.6
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1 